+++ config/webpack.js
@@ -363,7 +363,9 @@ module.exports = (env) => {
         // Notifications for item moves
         '$featureFlags.moveNotifications': JSON.stringify(true),
         // Item organizer
-        '$featureFlags.organizer': JSON.stringify(env.dev)
+        '$featureFlags.organizer': JSON.stringify(env.dev),
+        // Enable vendorengrams.xyz integration
+        '$featureFlags.vendorEngrams': JSON.stringify(true)
       }),
 
       new LodashModuleReplacementPlugin({